requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-crispy-forms==1.12.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.3.1
drf-extra-fields==3.1.1
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytz==2021.1
regex==2021.7.6
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
toml==0.10.2
tomli==1.1.0
uritemplate==3.0.1
urllib3==1.26.6
yapf==0.31.0
